Исправлены некоторые ошибки. Создан контроллер для создания турнира с участниками.

+++ back/tgbot/repository/src/main/java/dev/dubrovsky/repository/group/GroupMatchRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GroupMatchRepository extends JpaRepository<GroupMatch, Integer> {
+public interface GroupMatchRepository extends JpaRepository<GroupMatch, Long> {
