+++ back/tgbot/repository/src/main/java/dev/dubrovsky/repository/group/GroupStandingRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GroupStandingRepository extends JpaRepository<GroupStanding, Integer> {
+public interface GroupStandingRepository extends JpaRepository<GroupStanding, Long> {
