+++ back/tgbot/repository/src/main/java/dev/dubrovsky/repository/playoff/PlayoffMatchRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlayoffMatchRepository extends JpaRepository<PlayoffMatch, Integer> {
+public interface PlayoffMatchRepository extends JpaRepository<PlayoffMatch, Long> {
