+++ back/tgbot/repository/src/main/java/dev/dubrovsky/repository/standing/StandingRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StandingRepository extends JpaRepository<Standing, Integer> {
+public interface StandingRepository extends JpaRepository<Standing, Long> {
