+++ back/tgbot/repository/src/main/java/dev/dubrovsky/repository/tournament/TournamentRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TournamentRepository extends JpaRepository<Tournament, Integer> {
+public interface TournamentRepository extends JpaRepository<Tournament, Long> {
