+++ back/tgbot/repository/src/main/java/dev/dubrovsky/repository/user/UserStatisticsRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserStatisticsRepository extends JpaRepository<UserStatistics, Integer> {
+public interface UserStatisticsRepository extends JpaRepository<UserStatistics, Long> {
